Fixed variable type for window size.
See issue #103

+++ multicore_engine_core/include/glfw_wrapper/window.hpp
@@ -27,7 +27,7 @@ class window {
 	std::unique_ptr<window_callbacks> callbacks_;
 
 public:
-	window(const std::string& title, const glm::vec2& size);
+	window(const std::string& title, const glm::ivec2& size);
 	~window();
 };
